Integrazione di JavaDoc nella parte di codice 'Forecast'
- aggiunto il grassetto nei commenti
- aggiunti i link
- apportate alcune modifiche

+++ ancyb_diagnosticTool/src/main/java/it/univpm/ancyb_diagnosticTool/stats/AverageCurrentDirection.java
@@ -7,23 +7,52 @@ import it.univpm.ancyb_diagnosticTool.model.Forecast;
 import it.univpm.ancyb_diagnosticTool.model.ForecastObject;
 import it.univpm.ancyb_diagnosticTool.utilities.Time;
 
-
+/*
+ * <b>Statistica</b> che implementa l'interfaccia {@link it.univpm.ancyb_diagnosticTool.stats.StatsInterface}, 
+ * usata per calcolare il valore medio della direzione della corrente marina.
+ * La statistica è calcolata a partire dalla data e ora attuali, fino alla stessa ora di 'days' giorni dopo.
+ * 
+ * @implements StatsInterface
+ * @see it.univpm.ancyb_diagnosticTool.filters.StatsInterface
+ * 
+ * @author Manuele Silvestrini
+ */
 public class AverageCurrentDirection implements StatsInterface{
 	
 	private Forecast forecastForStats;
-	private int days;
+	private byte days;
 	private String stats = null;
 
+	/*
+	 * <b>Costruttore</b> della statistica.
+	 * @param forecastForStats Oggetto
+	 * {@link it.univpm.ancyb_diagnosticTool.model.Forecast#Forecast(ArrayList) Forecast}
+	 * in ingresso alla statistica.
+	 * @param days Parametro che contiene i giorni per cui si vuole esterdere la statistica.
+	 */
 	public AverageCurrentDirection(Forecast forecastForStats, byte days) {
 		this.forecastForStats = forecastForStats;
 		this.days = days;
 	}
 	
+	/*
+	 * @see it.univpm.ancyb_diagnosticTool.stats.StatsInterface#getDataForStats()
+	 */
 	@Override
 	public Object getDataForStats() {
 		return this.forecastForStats;
 	}
 
+	/*
+	 * <b>Metodo</b> che scorre tutti gli elementi dell'ArrayList contenuti nell'oggetto 
+	 * {@link it.univpm.ancyb_diagnosticTool.model.Forecast#Forecast(ArrayList) Forecast}
+	 * il quale tempo corrisponde con quello corrente, fino alla stessa ora di 'days' giorni successivi.
+	 * Nel mentre, da ogni oggetto viene prelevato il valore di direzione della corrente marina per calcolare la media,
+	 * che infine viene restituita al costruttore.
+	 * 
+	 * @throws StatsFailure
+	 * @see it.univpm.ancyb_diagnosticTool.stats.StatsInterface#computeStats()
+	 */
 	@Override
 	public void computeStats() throws StatsFailure{	
 
@@ -33,7 +62,6 @@ public class AverageCurrentDirection implements StatsInterface{
 		int totalHours = days*24;
 		int index = 0;
 		
-		//prendo i valori dell'elemento del tempo corrente
 		for(ForecastObject fobj : list) {
 						
 			if(fobj.getForecastTime().equals(Time.currentDateTime2())) {
@@ -43,7 +71,6 @@ public class AverageCurrentDirection implements StatsInterface{
 			else continue;
 		}
 		
-		//scorro i dati e sommo 
 		for(int i = 0; i < totalHours; i++) {
 			sum += list.get(index+i).getCurrentDirection();
 		}
@@ -52,7 +79,11 @@ public class AverageCurrentDirection implements StatsInterface{
 		
 		this.stats = String.format("%.2f", sum/totalHours);
 	}
-
+	
+	/*
+	 * @throws StatsFailure
+	 * @see it.univpm.ancyb_diagnosticTool.stats.StatsInterface#getStats()
+	 */
 	@Override
 	public String getStats() throws StatsFailure {
 		if ( stats == null) throw new StatsFailure("StatsFailure(AverageWaveHeight) --> please, first invoke the method 'computeStats()'.");
